Added .NET Foundation copyright notice.
