+++ src/UETools.Core/StreamExtensions.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace UETools.Core
 {
-    public static class StreamExtensions
+    public static partial class StreamExtensions
     {
         public static void ReadWholeBuf(this Stream stream, Span<byte> buf)
         {
@@ -89,7 +89,6 @@ namespace UETools.Core
             }
         }
 
-        // TODO: Copyrights
         public static int Read(this Stream stream, Memory<byte> buffer)
         {
             if (MemoryMarshal.TryGetArray<byte>(buffer, out var array))
@@ -111,27 +110,6 @@ namespace UETools.Core
                 }
             }
         }
-        public static void Write(this Stream stream, ReadOnlyMemory<byte> buffer)
-        {
-            if (MemoryMarshal.TryGetArray(buffer, out var array))
-            {
-                stream.Write(array.Array!, array.Offset, array.Count);
-            }
-            else
-            {
-                var sharedBuffer = ArrayPool<byte>.Shared.Rent(buffer.Length);
-                try
-                {
-                    buffer.Span.CopyTo(sharedBuffer);
-                    stream.Write(sharedBuffer, 0, buffer.Length);
-                }
-                finally
-                {
-                    ArrayPool<byte>.Shared.Return(sharedBuffer);
-                }
-            }
-        }
-
 #if NETSTANDARD2_0
         public static int Read(this Stream stream, Span<byte> buffer)
         {
@@ -158,57 +136,6 @@ namespace UETools.Core
             finally
             {
                 ArrayPool<byte>.Shared.Return(sharedBuffer);
-            }
-        }
-
-        public static ValueTask<int> ReadAsync(this Stream stream, Memory<byte> buffer, CancellationToken cancellationToken = default)
-        {
-            if (MemoryMarshal.TryGetArray<byte>(buffer, out var array))
-            {
-                return new ValueTask<int>(stream.ReadAsync(array.Array, array.Offset, array.Count, cancellationToken));
-            }
-            else
-            {
-                var sharedBuffer = ArrayPool<byte>.Shared.Rent(buffer.Length);
-                return FinishReadAsync(stream.ReadAsync(sharedBuffer, 0, buffer.Length, cancellationToken), sharedBuffer, buffer);
-
-                async ValueTask<int> FinishReadAsync(Task<int> readTask, byte[] localBuffer, Memory<byte> localDestination)
-                {
-                    try
-                    {
-                        var result = await readTask.ConfigureAwait(false);
-                        new Span<byte>(localBuffer, 0, result).CopyTo(localDestination.Span);
-                        return result;
-                    }
-                    finally
-                    {
-                        ArrayPool<byte>.Shared.Return(localBuffer);
-                    }
-                }
-            }
-        }
-        public static ValueTask WriteAsync(this Stream stream, ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken = default)
-        {
-            if (MemoryMarshal.TryGetArray(buffer, out var array))
-            {
-                return new ValueTask(stream.WriteAsync(array.Array, array.Offset, array.Count, cancellationToken));
-            }
-            else
-            {
-                var sharedBuffer = ArrayPool<byte>.Shared.Rent(buffer.Length);
-                buffer.Span.CopyTo(sharedBuffer);
-                return new ValueTask(FinishWriteAsync(stream.WriteAsync(sharedBuffer, 0, buffer.Length, cancellationToken), sharedBuffer));
-            }
-        }
-        private static async Task FinishWriteAsync(Task writeTask, byte[] localBuffer)
-        {
-            try
-            {
-                await writeTask.ConfigureAwait(false);
-            }
-            finally
-            {
-                ArrayPool<byte>.Shared.Return(localBuffer);
             }
         }
 #endif
